split tarballs into its own route file
