+++ routes/packages.js
@@ -4,7 +4,6 @@ const r = require('koa-router')()
 const path = require('path')
 const url = require('url')
 const packages = require('../lib/packages')
-const tarballs = require('../lib/tarballs')
 const config = require('../config')
 
 function addShaToPath (p, sha) {
@@ -45,33 +44,6 @@ r.get('/:name', function * () {
   this.set('ETag', pkg.etag)
   this.set('Cache-Control', `public, max-age=${config.cache.packageTTL}`)
   this.body = pkg
-})
-
-function * tarball () {
-  let {scope, name, filename, sha} = this.params
-  let tarball = yield tarballs.get(scope ? `${scope}/${name}` : name, filename, sha)
-  if (!tarball) {
-    this.status = 404
-    this.body = {error: 'no tarball found'}
-    return
-  }
-  this.set('Content-Length', tarball.size)
-  this.set('Cache-Control', `public, max-age=${config.cache.tarballTTL}`)
-  this.body = tarball.stream
-}
-
-// get package tarball with sha
-r.get('/:name/-/:filename/:sha', tarball)
-
-// get scoped package tarball with sha
-r.get('/:scope/:name/-/:filename/:sha', tarball)
-
-// get package tarball without sha
-r.get('/:name/-/:filename', function * () {
-  let {name, filename} = this.params
-  let ext = path.extname(filename)
-  filename = path.basename(filename, ext)
-  this.redirect(`/${name}/-/${filename}/a${ext}`)
 })
 
 module.exports = r
